requirements.txt
google-api-python-client
PyYAML
requests
